chatbot-app/backend/personalization/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
pydantic==2.12.4
python-multipart==0.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4